backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.8
tortoise-orm==0.21.3
aerich==0.7.2
asyncpg==0.29.0
authlib==1.3.1
atlassian-python-api==3.41.13
openai==1.35.3
anthropic==0.28.1
sqlalchemy[asyncio]==2.0.31
greenlet==3.0.3
